[SOIN] Remplace l’ancien champ texte
...en supprimant l’ancien qui encapsulait un input et en renommant le composant qui encapsule un dsfr-input
